+++ src/common/utils/kaltura-params.js
@@ -1,5 +1,5 @@
 //@flow
-import {Utils} from 'playkit-js'
+import {StreamType, Utils} from 'playkit-js'
 
 const PLAY_MANIFEST = 'playmanifest/';
 const PLAY_SESSION_ID = 'playSessionId=';
@@ -8,76 +8,106 @@ const CLIENT_TAG = 'clientTag=html5:v';
 
 /**
  * @param {Player} player - player
+ * @param {PartialKalturaPlayerOptionsObject} playerConfig - player config
  * @return {void}
  * @public
  */
-function handleSessionId(player: Player): void {
+function handleSessionId(player: Player, playerConfig: PartialKalturaPlayerOptionsObject): void {
   if (player.config.session && player.config.session.id) { // on change media
-    updateSessionId(player);
+    updateSessionId(player, playerConfig);
   } else { // on first playback
-    addSessionId(player);
+    addSessionId(playerConfig);
   }
 }
 
 /**
- * @param {Player} player - player
+ * @param {PartialKalturaPlayerOptionsObject} playerConfig - player config
  * @return {void}
  * @private
  */
-function addSessionId(player: Player): void {
+function addSessionId(playerConfig: PartialKalturaPlayerOptionsObject): void {
   let primaryGUID = Utils.Generator.guid();
   let secondGUID = Utils.Generator.guid();
-  player.sessionId = primaryGUID + ':' + secondGUID;
+  setSessionId(playerConfig, primaryGUID + ':' + secondGUID);
 }
 
 /**
  * @param {Player} player - player
+ * @param {PartialKalturaPlayerOptionsObject} playerConfig - player config
  * @return {void}
  * @private
  */
-function updateSessionId(player: Player): void {
+function updateSessionId(player: Player, playerConfig: PartialKalturaPlayerOptionsObject): void {
   let secondGuidInSessionIdRegex = /:((?:[a-z0-9]|-)*)/i;
   let secondGuidInSessionId = secondGuidInSessionIdRegex.exec(player.config.session.id);
   if (secondGuidInSessionId && secondGuidInSessionId[1]) {
-    player.sessionId = player.config.session.id.replace(secondGuidInSessionId[1], Utils.Generator.guid());
+    setSessionId(playerConfig, player.config.session.id.replace(secondGuidInSessionId[1], Utils.Generator.guid()));
   }
 }
 
 /**
- * @param {Object} source - source
+ * @param {PartialKalturaPlayerOptionsObject} playerConfig - player config
+ * @param {string} sessionId - the session id
+ * @return {void}
+ * @private
+ */
+function setSessionId(playerConfig: PartialKalturaPlayerOptionsObject, sessionId: string): void {
+  playerConfig.session = playerConfig.session || {};
+  playerConfig.session.id = sessionId;
+}
+
+/**
+ * @param {PKMediaSourceObject} source - PKMediaSourceObject
  * @param {string} sessionId - session id
  * @return {void}
  * @private
  */
-function updateSessionIdInUrl(source: Object = {}, sessionId: string): void {
-  let sessionIdInUrlRegex = new RegExp(PLAY_SESSION_ID + '((?:[a-z0-9]|-)*:(?:[a-z0-9]|-)*)', 'i');
-  let sessionIdInUrl = sessionIdInUrlRegex.exec(source.url);
-  if (sessionIdInUrl && sessionIdInUrl[1]) { // this url has session id (has already been played)
-    source.url = source.url.replace(sessionIdInUrl[1], sessionId);
-  } else {
-    let delimiter = source.url.indexOf('?') === -1 ? '?' : '&';
-    source.url += delimiter + PLAY_SESSION_ID + sessionId;
+function updateSessionIdInUrl(source: Object = {}, sessionId: ?string): void {
+  if (sessionId) {
+    let sessionIdInUrlRegex = new RegExp(PLAY_SESSION_ID + '((?:[a-z0-9]|-)*:(?:[a-z0-9]|-)*)', 'i');
+    let sessionIdInUrl = sessionIdInUrlRegex.exec(source.url);
+    if (sessionIdInUrl && sessionIdInUrl[1]) { // this url has session id (has already been played)
+      source.url = source.url.replace(sessionIdInUrl[1], sessionId);
+    } else {
+      let delimiter = source.url.indexOf('?') === -1 ? '?' : '&';
+      source.url += delimiter + PLAY_SESSION_ID + sessionId;
+    }
   }
 }
 
 /**
- * @param {Object} source - source
+ * @return {string} - The referrer
+ * @private
+ */
+function getReferrer(): string {
+  let referrer;
+  try {
+    referrer = window.parent.document.URL;
+  } catch (e) { // unfriendly iframe
+    referrer = document.referrer;
+  }
+  return referrer;
+}
+
+/**
+ * @param {PKMediaSourceObject} source - source
  * @return {void}
  * @private
  */
-function addReferrer(source: Object) {
+function addReferrer(source: PKMediaSourceObject): void {
   if (source.url.indexOf(REFERRER) === -1) {
     let delimiter = source.url.indexOf('?') === -1 ? '?' : '&';
-    source.url += delimiter + REFERRER + btoa(document.referrer || document.URL);
+    let referrer = btoa(getReferrer().substr(0, 1000));
+    source.url += delimiter + REFERRER + referrer;
   }
 }
 
 /**
- * @param {Object} source - source
+ * @param {PKMediaSourceObject} source - source
  * @return {void}
  * @private
  */
-function addClientTag(source: Object) {
+function addClientTag(source: PKMediaSourceObject): void {
   if (source.url.indexOf(CLIENT_TAG) === -1) {
     let delimiter = source.url.indexOf('?') === -1 ? '?' : '&';
     source.url += delimiter + CLIENT_TAG + __VERSION__;
@@ -85,22 +115,27 @@ function addClientTag(source: Object) {
 }
 
 /**
- * @param {Object} sources - player sources
+ * Adding Kaltura specific params to player config and player sources.
  * @param {Player} player - player
+ * @param {PartialKalturaPlayerOptionsObject} playerConfig - player config
  * @return {void}
  * @private
  */
-function addKalturaParams(sources: Object, player: Player): void {
-  handleSessionId(player);
-  Object.keys(sources).forEach((key) => {
-    sources[key].forEach((source) => {
-      if (typeof source.url === 'string' && source.url.toLowerCase().indexOf(PLAY_MANIFEST) > -1) {
-        updateSessionIdInUrl(source, player.config.session.id);
-        addReferrer(source);
-        addClientTag(source);
-      }
-    });
+function addKalturaParams(player: Player, playerConfig: PartialKalturaPlayerOptionsObject): void {
+  handleSessionId(player, playerConfig);
+  const sources = playerConfig.sources;
+  Object.values(StreamType).forEach(key => {
+    // $FlowFixMe
+    if (sources[key]) {
+      sources[key].forEach((source) => {
+        if (typeof source.url === 'string' && source.url.toLowerCase().indexOf(PLAY_MANIFEST) > -1 && !source.localSource) {
+          updateSessionIdInUrl(source, playerConfig.session && playerConfig.session.id);
+          addReferrer(source);
+          addClientTag(source);
+        }
+      });
+    }
   });
 }
 
-export {addKalturaParams, handleSessionId, updateSessionIdInUrl, addReferrer, addClientTag}
+export {addKalturaParams, handleSessionId, updateSessionIdInUrl, getReferrer, addReferrer, addClientTag}