+++ src/common/utils/thumbs.js
@@ -7,17 +7,17 @@ const TEMPLATE: string = '{{thumbnailUrl}}/width/{{width}}/vid_slices/{{slices}}
 
 /**
  * Builds thumbnail slices url for the ui.
- * @param {Object} data - The prover data.
- * @param {Object} uiConfig - The ui config.
+ * @param {ProviderMediaConfigObject} mediaConfig - The prover media config.
+ * @param {SeekbarConfig?} seekbarConfig - The seek bar config.
  * @returns {string} - The thumbnail slices url.
  */
-export function getThumbSlicesUrl(data: Object, uiConfig: Object = {}): string {
+export function getThumbSlicesUrl(mediaConfig: ProviderMediaConfigObject, seekbarConfig?: SeekbarConfig): string {
   try {
     const model: Object = {
-      thumbnailUrl: data.metadata.poster,
-      ks: data.session.ks,
-      width: uiConfig.thumbsWidth || DEFAULT_THUMBS_WIDTH,
-      slices: uiConfig.thumbsSlices || DEFAULT_THUMBS_SLICES,
+      thumbnailUrl: mediaConfig.sources.poster,
+      ks: mediaConfig.session.ks,
+      width: (seekbarConfig && seekbarConfig.thumbsWidth) || DEFAULT_THUMBS_WIDTH,
+      slices: (seekbarConfig && seekbarConfig.thumbsSlices) || DEFAULT_THUMBS_SLICES,
     };
     return evaluate(TEMPLATE, model);
   } catch (e) {