+++ test/src/common/utils/thumbs.spec.js
@@ -2,7 +2,7 @@ import {getThumbSlicesUrl} from '../../../../src/common/utils/thumbs'
 
 describe('getThumbSlicesUrl', function () {
   const fakeData = {
-    metadata: {
+    sources: {
       poster: "//my-thumb-service.com/p/1/entry_id/2/version/3"
     },
     session: {
@@ -17,13 +17,13 @@ describe('getThumbSlicesUrl', function () {
 
   it('should get thumbnail slices url with default params', function () {
     getThumbSlicesUrl(fakeData).should.equals(
-      `${fakeData.metadata.poster}/width/164/vid_slices/100/ks/${fakeData.session.ks}`
+      `${fakeData.sources.poster}/width/164/vid_slices/100/ks/${fakeData.session.ks}`
     );
   });
 
   it('should get thumbnail slices url with the custom params', function () {
     getThumbSlicesUrl(fakeData, fakeUIConfig).should.equals(
-      `${fakeData.metadata.poster}/width/100/vid_slices/200/ks/${fakeData.session.ks}`
+      `${fakeData.sources.poster}/width/100/vid_slices/200/ks/${fakeData.session.ks}`
     );
   });
 });